+++ packages/utils/src/index.js
@@ -1,5 +1,6 @@
 export * from './arrays.js';
 export * from './browser.js';
+export * from './cache.js';
 export * from './cloning.js';
 export * from './colors.js';
 export * from './crypto.js';